refactor: move user relationships to new resource model
