+++ rgdps/resources/user.py
@@ -89,11 +89,6 @@ class UserPrivacySetting(IntEnum):
     PRIVATE = 2
 
 
-class UserRelationshipType(IntEnum):
-    FRIEND = 0
-    BLOCKED = 1
-
-
 class UserPrivilegeLevel(IntEnumStringMixin, IntEnum):
     """Enum for determining whether a user should be displayed as a
     moderator, elder moderator, or neither.
